samples: lz4: Measure duration of a firmware image decompression
Port LZ4 sample to nRF54LM20 to measure performance (speed) of
a firmware image decompression. This sample uses an ESB PRX image
from another prestudy as the test data. The measurements are performed
executing the code from RAM and RRAM. It is expected that RAM execution
would be more efficient in real DFU scenario if it could be performed
in parallel with RRAM writing of the previous block.

The memory requirements of LZ4 render this algorithm unfitting nRF54L
series. However its speed prooves that other algorithms like nrv2e or
Exomizer with lower memory requirements and comparable speed could
fit DFU well.

Signed-off-by: Hubert Miś <hubert.mis@nordicsemi.no>